requirements.txt
fastapi
uvicorn
redis
debugpy
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-multipart
python-jose[cryptography]
alembic